requirements.txt
dj-database-url==0.3.0
Django==1.8.1
django-widget-tweaks==1.3
gunicorn==19.3.0
psycopg2==2.6
whitenoise==1.0.6
